Move newPipeline() out to be an exported function
