+++ sdk/storage/storage-blob/src/index.ts
@@ -8,7 +8,6 @@ import * as Models from "../src/generated/lib/models";
 export * from "./BlobServiceClient";
 export * from "./ContainerClient";
 // ordering of these clients is important as there are circular dependencies
-export * from "./StorageClient";
 export * from "./BlobClient";
 export * from "./AppendBlobClient";
 export * from "./BlockBlobClient";
