+++ sdk/storage/storage-blob/test/utils/index.ts
@@ -4,9 +4,9 @@ import * as path from "path";
 
 import { SharedKeyCredential } from "../../src/credentials/SharedKeyCredential";
 import { BlobServiceClient } from "../../src/BlobServiceClient";
-import { StorageClient } from "../../src/StorageClient";
 import { getUniqueName } from "./testutils.common";
 import * as dotenv from "dotenv";
+import { newPipeline } from "../../src/Pipeline";
 dotenv.config({ path: "../.env" });
 
 export * from "./testutils.common";
@@ -26,7 +26,7 @@ export function getGenericBSU(accountType: string, accountNameSuffix: string = "
   }
 
   const credentials = new SharedKeyCredential(accountName, accountKey);
-  const pipeline = StorageClient.newPipeline(credentials, {
+  const pipeline = newPipeline(credentials, {
     // Enable logger when debugging
     // logger: new ConsoleHttpPipelineLogger(HttpPipelineLogLevel.INFO)
   });
